# server/docs/swagger.yaml
basePath: /
definitions:
  art.createRequest:
    properties:
      description:
        example: Deu muito trabalho!
        type: string
      image_url:
        example: https://exemplo.com/arte.jpg
        type: string
      title:
        example: Fanart do Goku
        maxLength: 100
        minLength: 2
        type: string
    required:
    - image_url
    - title
    type: object
  art.updateRequest:
    properties:
      description:
        type: string
      title:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
  domain.Art:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      likes_count:
        type: integer
      profile_id:
        type: string
      title:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - image_url
    - profile_id
    - title
    type: object
...
info:
  contact: {}
  description: API central da plataforma GeekVerse (Streaming, Leitura e Galeria de
    Artes).
  title: GeekVerse API
  version: "1.0"
paths:
  /arts:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Art'
            type: array
      tags:
      - Arts
    post:
      parameters:
      - description: Dados da Arte
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/art.createRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Art'
      security:
      - BearerAuth: []
      tags:
      - Arts
  /arts/{id}:
    delete:
      parameters:
      - description: ID da Arte
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Arts
    get:
      parameters:
      - description: ID da Arte
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Art'
      tags:
      - Arts
    patch:
      parameters:
      - description: ID da Arte
        in: path
        name: id
        required: true
        type: integer
      - description: Campos para atualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/art.updateRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Arts
  /arts/{id}/likes:
    delete:
      parameters:
      - description: ID da Arte
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Arts
    post:
      parameters:
      - description: ID da Arte
        in: path
        name: id
        required: true
        type: integer
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Arts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"